+++ packages/query-devtools/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @tanstack/query-devtools
 
+## 5.96.2
+
 ## 5.96.1
 
 ## 5.96.0